move App.tsx and App.scss into components/
